+++ src/components/ExportCSV/CSVPreview/CSVPreview.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CSVPreview = () => {
+  return (
+    <div>
+      <p>CSV Preview</p>
+    </div>
+  );
+};
+
+export default CSVPreview;
